Import chromosomal position information also for constant regions. Support for IMGT constant region lists. Bug fixes for NCBIm38 build scripts. Always check for existence of ".my.cnf" at start. Remove temporary files from repo. Capitalize binomial species names.

+++ build_library_ncbim38.sh
@@ -2,20 +2,26 @@ export CURR_DATE=$( date --iso-8601 )
 export LIBRARY="library_mouse_ncbim38"
 export DIR_IGDATA_DB="igdata/database"
 
+if [[ ! -e $HOME/.my.cnf ]]; then
+	echo "Could not find MySQL/MariaDB local config file \"$HOME/.my.cnf\". This file is required for providing"
+	echo "authentication information to the script. Aborting!"
+	exit 1;
+fi
+
 mkdir -p $DIR_IGDATA_DB
 
 # set up library schemes
 mysql --defaults-file=$HOME/.my.cnf --defaults-group-suffix=_igdb -e "CREATE SCHEMA IF NOT EXISTS ${LIBRARY};"
-mysql --defaults-file=$HOME/.my.cnf --defaults-group-suffix=_igdb --database=${LIBRARY}  < igdb_library.sql
+mysql --defaults-file=$HOME/.my.cnf --defaults-group-suffix=_igdb --database=${LIBRARY} < igdb_library.sql
 
 # process mouse VDJ segment data 
 #
 cd VDJ_segments
 
-for file in data_ncbim38_mouse_2015-05-14/mouse*.n*; 
+for file in data_mouse_ncbim38_2015-05-14/mouse*.n*; 
 	do cp -i "${file}" "${file/mouse_gl/mouse_ncbim38_gl}";		# Copy and rename"
 done
-mv data_ncbim38_mouse_2015-05-14/mouse_ncbim38*.n* ../${DIR_IGDATA_DB}
+mv -v data_mouse_ncbim38_2015-05-14/mouse_ncbim38*.n* ../${DIR_IGDATA_DB}
 
 echo "[OK] Setting up database for NCBI m38 mouse VDJ."
 ./build_VDJ_db.pl \
@@ -33,7 +39,7 @@ echo "[OK] Setting up database for NCBI m38 mouse VDJ."
 cd ../constant_segments/
 
 echo "[OK] Setting up database for mouse constant."
-mv -v mouse_ncbim38_gl_C.fasta ../${DIR_IGDATA_DB}/mouse_ncbim38_gl_C
+cp -v mouse_ncbim38_gl_C.fasta ../${DIR_IGDATA_DB}/mouse_ncbim38_gl_C
 makeblastdb -dbtype nucl -in ../${DIR_IGDATA_DB}/mouse_ncbim38_gl_C
 mv -v ../${DIR_IGDATA_DB}/mouse_ncbim38_gl_C ../${DIR_IGDATA_DB}/mouse_ncbim38_gl_C.fasta
 ./build_constant_db.pl -mysql_group mysql_igdb -lib ${LIBRARY} -sp mouse -fasta ../${DIR_IGDATA_DB}/mouse_ncbim38_gl_C.fasta
@@ -42,6 +48,7 @@ mv -v ../${DIR_IGDATA_DB}/mouse_ncbim38_gl_C ../${DIR_IGDATA_DB}/mouse_ncbim38_g
 #
 cd ../tags
 ./process_tags.pl
+
 cd ..
 
 # Insert remaining tables into the database
